requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.3
cffi==2.0.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==47.0.0
fastapi==0.136.1
h11==0.16.0
idna==3.13
invoke==3.0.3
kombu==5.6.2
packaging==26.2
paramiko==4.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyMySQL==1.1.2
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uvicorn==0.46.0
vine==5.1.0
wcwidth==0.6.0
